Detection of enableHelm
The deletion of inactive helm releases as currently implemented is very aggressive.

In the civils project we occasionally have the need to refer PRs within other PRs for build purposes.
Also, testers may take some time before they get to a certain PR.

Having the deletion of inactive helm releases behave so aggressively is getting in the way of both development and testing, so we need to start using the "enableHelm" label on our PRs to discern what can be safely deleted and what should be kept.

This change introduces a check against the enableHelm label on github to prevent deletion of pods that for whatever reason the developer or tester explicitly decided to keep.